getSafeTypes, more test data stuff
