+++ Hosting/Abstractions/src/Dosaic.Hosting.Abstractions/Extensions/AssemblyExtensions.cs
@@ -4,9 +4,21 @@ namespace Dosaic.Hosting.Abstractions.Extensions
 {
     public static class AssemblyExtensions
     {
+        private static Type[] GetTypesSafe(Assembly assembly)
+        {
+            try
+            {
+                return assembly.GetTypes();
+            }
+            catch
+            {
+                return [];
+            }
+        }
+
         public static bool HasType(this Assembly assembly, Predicate<Type> typePredicate)
         {
-            return assembly.GetTypes().Any(t => typePredicate(t));
+            return GetTypesSafe(assembly).Any(t => typePredicate(t));
         }
 
         public static IList<Type> GetTypes(this IEnumerable<Assembly> assemblies, Predicate<Type> typePredicate = null)
@@ -16,7 +28,7 @@ namespace Dosaic.Hosting.Abstractions.Extensions
 
         public static IList<Type> GetTypes(this Assembly assembly, Predicate<Type> typePredicate = null)
         {
-            return assembly.GetTypes().Where(x => typePredicate == null || typePredicate(x)).ToList();
+            return GetTypesSafe(assembly).Where(x => typePredicate == null || typePredicate(x)).ToList();
         }
     }
 }
